+++ service/createTask.go
@@ -6,11 +6,6 @@ import (
 	"go.uber.org/zap"
 )
 
-type TaskService struct {
-	repo   *repository.TaskRepository
-	logger *zap.Logger
-}
-
 func NewTaskService(repo *repository.TaskRepository, logger *zap.Logger) *TaskService {
 	return &TaskService{repo: repo, logger: logger}
 }
